Figure 5
Figure 5 - Distribution of inferred miRNA family losses on the placental phylogeny used in this study. The position of miRNA family losses are denoted with circles and the colour of these circles represent each individual miRNA gene. miRNA genes that were 'absent' in species marked with an asterisk were found to be 'present' in our alignment-based analyses (see Materials and Methods).